Committing before creating localposition branch.
This commit represents the latest version of the model, where we
override the __eq__ methods of Atom and Bond classes, and __eq__
and __hash__ methods of Pattern class to use Counter to produce
data for clustering. In this model, atom id is the id along the
entire backbone. However it may be simpler and more logical to
use the local id, i.e. position along the corresponding backbone
segment. This will be implemented in the localposition branch.